BookStore/WebApi/Application/GenreOperations/Commands/DeleteGenre/DeleteGenreCommand.cs
using AutoMapper;
using Microsoft.EntityFrameworkCore;
using WebApi.DBOperations;

namespace WebApi.Application.GenreOperations.Commands.DeleteGenre
{
    public class DeleteGenreCommand
    {
        private readonly IBookStoreDbContext _context;
        public int Id {get; set; }

        public DeleteGenreCommand(IBookStoreDbContext context)
        {
            _context = context;
        }

        public void Handle()
        {
            var genre = _context.Genres.SingleOrDefault(x => x.Id == Id);
            if(genre is null)
                throw new InvalidOperationException("The genre is not found");
            if(_context.Books.Any(x => x.GenreId == Id))
                throw new InvalidOperationException("The genre is using by any book");
            _context.Genres.Remove(genre);
            _context.SaveChanges();
        }
    }
}